Additional accessors; version bumped to 1.0.5
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_3_0/madman/Rpacks/kebabs@100131 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/GappyPairC.cpp
@@ -2277,7 +2277,7 @@ static bool getIndexMap(ByteStringVector x, int sizeX, IntegerVector selX, ByteS
                 ks_mergesort(gappy, noOfEntries, featureIndices, 0);
 
                 // write col index to hash
-                for (int i=0; i < *numUsedFeatures; i++)
+                for (int i=0; i < (int) *numUsedFeatures; i++)
                 {
                     if (i % USER_INTERRUPT_LIMIT == 0)
                         R_CheckUserInterrupt();
@@ -2652,7 +2652,7 @@ void getERDGappy(NumericMatrix erd, ByteStringVector x, int sizeX, IntegerVector
 
             iX = selX[i];
 
-            for (j = 0; j < numUsedFeatures; j++)
+            for (j = 0; j < (int) numUsedFeatures; j++)
             {
                 if (erd(i,j) > 0)
                     erd(i,j) = erd(i,j) / normValues[i];
